refactor: rename statusUtils
